remove the new http way and mantain the compatibility
